Make WikiFSEngine Linux-buildable: conditional ACP dependency (#754, #780)
The swift-acp  product uses ACPProcessManager and os.log — both
macOS-only. This made WikiFSEngine unbuildable on Linux, blocking the
linux-swift CI job.

Changes:
- Package.swift: make the  product a macOS-only dependency of
  WikiFSEngine (condition: .when(platforms: [.macOS])). ACPModel stays
  unconditional (pure model types, portable). WikiFSEngine is now a
  macOS-only dep of WikiFSCoreTests, WikiFSAppTests, and WikiFS.
- Extract portable value types (SessionUsage, PendingPermission,
  PermissionPolicy) from ACPBackend.swift / ACPPermissions.swift into
  ACPModelValueTypes.swift — these are pure structs/enums with no ACP
  dependency, used by the portable queue system.
- Guard the 3 ACP-importing source files (ACPBackend, ACPPermissions,
  ACPProviderModelProbe) with #if os(macOS).
- Guard the agent-execution cluster (AgentLauncher, AgentOperationRunner,
  ACPExtractionClient, MessageSummarizer, QuotaFallbackCoordinator,
  WikiSession, SessionManager, ExtractionCoordinator, SpawnModelGuard)
  with #if os(macOS) — these are deeply coupled to ACPBackend.
- Guard the ACPBackend: PermissionResolving conformance extension and
  AgentBackendFactory.makeBackend's ACPBackend construction with
  #if os(macOS) (makeBackend fatalError's on Linux — it's only called from
  macOS code paths).
- Add LinuxStubBackend.swift: a no-op AgentBackend conformance for Linux
  type resolution (never driven — all AgentLauncher tests are guarded).

Validation:
- swift build passes on macOS (no warnings).
- swift test --filter WikiFSCoreTests: 2194 tests in 163 suites, all green.
- Linux path validated by CI.
- Refs #754, #780.
